Linear interpolation extraction works (see message)
Interpreting the "shifted grid" whose nodes are centers of the cells, and taking as B and C
the maximum traversability cost of one of the 4 original overlapping cells, LERP works.
However, I doubt that taking the max is optimal: the average should have been better, but it
doesn't work for some reason.